# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Montor-ing API",
        "description": "Dokumentasi API untuk aplikasi Montor-ing (Inventory & Sales Monitoring).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Server API Utama"
        }
    ],
    "paths": {
...
        "/products/low-stock-alert": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Alert Stok Menipis (Stok <= 5)",
                "operationId": "1fd97685a231ff33c08fa1a039acc9c7",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit jumlah data",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data produk dengan stok menipis berhasil diambil.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data produk dengan stok menipis berhasil diambil."
                                        },
                                        "count_returned": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "total_data": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "sku": {
                                                        "type": "string",
                                                        "example": "PROD001"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Produk 1"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "example": 10000
                                                    },
                                                    "stock": {
                                                        "type": "integer",
                                                        "example": 3
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/products/{id}": {
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update Produk",
                "operationId": "2ce632b9d1c0c445557821d09add7ec2",
                "parameters": [
                    {
                        "name": "X-Database-Name",
                        "in": "header",
                        "description": "Nama Database",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Produk",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nama Produk",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Harga Produk",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produk berhasil diupdate.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Produk berhasil diupdate."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "parameters": {
            "X-Database-Name": {
                "name": "X-Database-Name",
                "in": "header",
                "description": "Nama database yang digunakan",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dari endpoint /login. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}